requirements.txt
boto3
psycopg2-binary
prometheus_client
tenacity
